:hammer: move gdrive images baking to its own file
